fix: solve crash when quitting the app
```
root.after(1000 // FRAME_PER_SECONDS)
self.tk.call('after', ms)
RuntimeError: main thread is not in main loop
```

Issue is that the `root.after` call being used in a separate thread, which
conflicts with the main Tkinter event loop. To avoid this issue, we should
not use root.after in a separate thread. Instead, we'll rely on Pygame's own
timing and rendering loop.

Signed-off-by: Lionel Hubaut <lionel.hubaut@dekimo.com>