refactor: big refactor FTW
refactor is the answer of the universe.

More seriously, I decoupled UI from the main application. The main app
will observe the change from the backend and will be notified when it
needs to refresh the display (observer pattern)

I also split the logic in different files for a clear separation of
concerns.

Signed-off-by: Lionel Hubaut <lionel.hubaut@dekimo.com>