chore(lint): fix some left over linting issue
src/pyvision/device (not device_ext) is the module that will be build
during the hatch env setup process.
In order for pylance to correctly lint a custom c module we need to
expose the types of the functions that the module exposes. That's the
purpose of the `src/pyvision/device.pyi` file which describe the public
methods and type hint that the extension exposes.

pylance is pyright under the hood, but directly integrated has an
extension.
Configured in strict mode to enforce good code quality.