fix(fps): wrong calculation
FPS were computed over the total number of frames by the total of time
spent. Which results in a overall smooth value overtime rather than the
instant fps rate. This is bad because lost frames would not result in a
significant drop in the counter, at least not immediate.

Now a moving average is calculated over the delta times over the last N
frames, which makes the computation more dynamic.

Signed-off-by: Lionel H <lionel.hubaut@dekimo.com>