refactor(mvc): introduce MVC architecture
Not an easy task :p

I mainly focused on the pygame frame and rendering processed frame into
the pygame canvas. This part is working well.

Camera switching logic is also done.

I think I will skip the brightness and contrast for now, because it does
not bring much benefit, the camera itself calibrate quite well with the
environment

Signed-off-by: Lionel H <lionel.hubaut@dekimo.com>