feat: make the stream handler a threaded class again
We only care about the last frame, we don't need to maintain a buffer of
image on the stream side. By using a threaded class, we are sure that
the frame presented by the stream model to the controller will always be
the latest frame that the model was able to read.

I checked using RTSP and without this I saw some buffering occuring.

Signed-off-by: Lionel H <lionel.hubaut@dekimo.com>