fix: the build extension was failing
It seems that recent change in the cpp file corrupt the build. The cause
was likely because the python was building the project with a missing
dependency to Ole32 and OleAuth32, strange that this was working before.
I only refactor code, so I don't really now why it suddenly start
failing.

I took this opportunity to also refactory setup.py to automatically
parse depencencies from the toml file rather than to duplicate
everything in a requirements.txt file.

Signed-off-by: Lionel H <lionel.hubaut@dekimo.com>