fix(fps): apply staturation for the fps counter
We could have a scenario where we can have an error between two reading,
leading to a potentially very small delta_time, resulting in a very high
fps value, which will obviously cause side effect for further
calculation since we are using a rolling window.
By check the bound, we ensure that the delta_time will remain between a
controlled value that would not cause big spike in fps calculation.

Signed-off-by: Lionel H <lionel.hubaut@dekimo.com>